# Optimized GrpcClientTest#serverStream