GrpcClientTest failed occasionally, see https://github.com/apache/dubbo/actions/runs/13621947200/job/38072959716
```
Tests run: 5, Failures: 1, Errors: 0, Skipped: 0, Time elapsed: 15.441 s <<< FAILURE! - in org.apache.dubbo.sample.tri.grpc.GrpcClientTest
clientSendLargeSizeHeader(org.apache.dubbo.sample.tri.grpc.GrpcClientTest)  Time elapsed: 1.108 s
unaryGreeter(org.apache.dubbo.sample.tri.grpc.GrpcClientTest)  Time elapsed: 0.325 s
stream(org.apache.dubbo.sample.tri.grpc.GrpcClientTest)  Time elapsed: 0.202 s
attachmentTest(org.apache.dubbo.sample.tri.grpc.GrpcClientTest)  Time elapsed: 0.023 s
serverStream(org.apache.dubbo.sample.tri.grpc.GrpcClientTest)  Time elapsed: 12.004 s  <<< FAILURE!
java.lang.AssertionError
	at org.junit.Assert.fail(Assert.java:87)
	at org.junit.Assert.assertTrue(Assert.java:42)
	at org.junit.Assert.assertTrue(Assert.java:53)
	at org.apache.dubbo.sample.tri.grpc.GrpcClientTest.serverStream(GrpcClientTest.java:95)
```
it might be meaningless that calling latch.countDown only when diff less than 50ms, 
```
                    long diff = Math.abs(now - lastTimestamp - 1000);
                    System.out.println(diff);
                    if (diff < 50) {
                        latch.countDown();
                    }
```