@zrlw 2025-03-05
i tested it on my own windows (4c) for a while, sometimes diff is bigger than 50ms,
![GrpcClientTest](https://github.com/user-attachments/assets/c3c8c296-a1c8-4900-bdaf-68b49e523dec)